requirements.txt
Django==3.1.2
djangorestframework==3.12.1
psycopg2==2.8.6
pytest==6.1.1
requests==2.24.0